Minor simulation and control enhancements
Added a margin to the default ChargeDisplayerRadioButton style, so the mouse is placing the center of the control.
Added a termination distance, so when a movable particle gets 1000000 points away form point [0,0], it gets automatically terminated.
If available, charge is auto-selected after AllowChargeSelection is set to true again.
Added pausing possibilities.